week_12_finance_research_agent/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pydantic>=2.7.0
tqdm>=4.66.0
python-dotenv>=1.0.1
reportlab>=4.2.2
fal-client>=0.5.9
requests>=2.32.3
openai>=1.45.0
matplotlib>=3.9.0
